Rename AgencyStore -> AgencyComponentStore
